Move AspNetCore proxies into a Proxies subfolder/namespace
