+++ tracer/src/Datadog.Trace/ClrProfiler/AutoInstrumentation/AspNetCore/Proxies/IAttributeRouteInfo.cs
@@ -1,4 +1,4 @@
-// <copyright file="IContext.cs" company="Datadog">
+// <copyright file="IAttributeRouteInfo.cs" company="Datadog">
 // Unless explicitly stated otherwise all files in this repository are licensed under the Apache 2 License.
 // This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2017 Datadog, Inc.
 // </copyright>
@@ -6,11 +6,11 @@
 #if NETFRAMEWORK
 #nullable enable
 
-namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.AspNetCore
+namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.AspNetCore.Proxies
 {
-    internal interface IContext
+    internal interface IAttributeRouteInfo
     {
-        IHttpContext HttpContext { get; }
+        string? Template { get; }
     }
 }
 #endif
