+++ tracer/src/Datadog.Trace/ClrProfiler/AutoInstrumentation/AspNetCore/Proxies/IContext.cs
@@ -1,19 +1,16 @@
-// <copyright file="IConnectionInfo.cs" company="Datadog">
+// <copyright file="IContext.cs" company="Datadog">
 // Unless explicitly stated otherwise all files in this repository are licensed under the Apache 2 License.
 // This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2017 Datadog, Inc.
 // </copyright>
 
 #if NETFRAMEWORK
 #nullable enable
-using System.Net;
 
-namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.AspNetCore
+namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.AspNetCore.Proxies
 {
-    internal interface IConnectionInfo
+    internal interface IContext
     {
-        IPAddress? RemoteIpAddress { get; }
-
-        int RemotePort { get; }
+        IHttpContext HttpContext { get; }
     }
 }
 #endif
