+++ tracer/src/Datadog.Trace/ClrProfiler/AutoInstrumentation/AspNetCore/Proxies/IQueryCollection.cs
@@ -1,4 +1,4 @@
-// <copyright file="IRequestCookieCollection.cs" company="Datadog">
+// <copyright file="IQueryCollection.cs" company="Datadog">
 // Unless explicitly stated otherwise all files in this repository are licensed under the Apache 2 License.
 // This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2017 Datadog, Inc.
 // </copyright>
@@ -7,14 +7,18 @@
 #nullable enable
 
 using System.Collections.Generic;
+using Datadog.Trace.DuckTyping;
 
-namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.AspNetCore
+namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.AspNetCore.Proxies
 {
-    internal interface IRequestCookieCollection : IEnumerable<KeyValuePair<string, string>>
+    internal interface IQueryCollection
     {
         int Count { get; }
 
         ICollection<string> Keys { get; }
+
+        [Duck(Name = "get_Item")]
+        string GetItemAsString(string key);
     }
 }
 #endif
