+++ tracer/src/Datadog.Trace/ClrProfiler/AutoInstrumentation/AspNetCore/Proxies/IResourceInvoker.cs
@@ -1,4 +1,4 @@
-// <copyright file="TemplateSegmentStruct.cs" company="Datadog">
+// <copyright file="IResourceInvoker.cs" company="Datadog">
 // Unless explicitly stated otherwise all files in this repository are licensed under the Apache 2 License.
 // This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2017 Datadog, Inc.
 // </copyright>
@@ -6,15 +6,14 @@
 #if NETFRAMEWORK
 #nullable enable
 
-using System.Collections;
 using Datadog.Trace.DuckTyping;
 
-namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.AspNetCore
+namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.AspNetCore.Proxies
 {
-    [DuckCopy]
-    internal struct TemplateSegmentStruct
+    internal interface IResourceInvoker
     {
-        public IEnumerable Parts;
+        [DuckField(Name = "_actionContext")]
+        IActionContext ActionContext { get; }
     }
 }
 #endif
