+++ tracer/src/Datadog.Trace/ClrProfiler/AutoInstrumentation/AspNetCore/Proxies/IRouteData.cs
@@ -1,4 +1,4 @@
-// <copyright file="IRoutingFeature.cs" company="Datadog">
+// <copyright file="IRouteData.cs" company="Datadog">
 // Unless explicitly stated otherwise all files in this repository are licensed under the Apache 2 License.
 // This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2017 Datadog, Inc.
 // </copyright>
@@ -6,13 +6,16 @@
 #if NETFRAMEWORK
 #nullable enable
 
+using System.Collections.Generic;
 using Datadog.Trace.DuckTyping;
 
-namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.AspNetCore
+namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.AspNetCore.Proxies
 {
-    internal interface IRoutingFeature
+    internal interface IRouteData : IDuckType
     {
-        IRouteData RouteData { get; }
+        IEnumerable<object> Routers { get; }
+
+        IDictionary<string, object> Values { get; }
     }
 }
 #endif
