+++ tracer/src/Datadog.Trace/ClrProfiler/AutoInstrumentation/AspNetCore/Proxies/IRouteTemplate.cs
@@ -1,4 +1,4 @@
-// <copyright file="RouteBaseStruct.cs" company="Datadog">
+// <copyright file="IRouteTemplate.cs" company="Datadog">
 // Unless explicitly stated otherwise all files in this repository are licensed under the Apache 2 License.
 // This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2017 Datadog, Inc.
 // </copyright>
@@ -6,14 +6,15 @@
 #if NETFRAMEWORK
 #nullable enable
 
-using Datadog.Trace.DuckTyping;
+using System.Collections;
 
-namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.AspNetCore
+namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.AspNetCore.Proxies
 {
-    [DuckCopy]
-    internal struct RouteBaseStruct
+    internal interface IRouteTemplate
     {
-        public IRouteTemplate ParsedTemplate;
+        string? TemplateText { get; }
+
+        IEnumerable Segments { get; }
     }
 }
 #endif
