+++ tracer/src/Datadog.Trace/ClrProfiler/AutoInstrumentation/AspNetCore/Proxies/RouteBaseStruct.cs
@@ -1,4 +1,4 @@
-// <copyright file="IAttributeRouteInfo.cs" company="Datadog">
+// <copyright file="RouteBaseStruct.cs" company="Datadog">
 // Unless explicitly stated otherwise all files in this repository are licensed under the Apache 2 License.
 // This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2017 Datadog, Inc.
 // </copyright>
@@ -6,11 +6,14 @@
 #if NETFRAMEWORK
 #nullable enable
 
-namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.AspNetCore
+using Datadog.Trace.DuckTyping;
+
+namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.AspNetCore.Proxies
 {
-    internal interface IAttributeRouteInfo
+    [DuckCopy]
+    internal struct RouteBaseStruct
     {
-        string? Template { get; }
+        public IRouteTemplate ParsedTemplate;
     }
 }
 #endif
