+++ tracer/src/Datadog.Trace/ClrProfiler/AutoInstrumentation/AspNetCore/Proxies/TemplatePartStruct.cs
@@ -8,7 +8,7 @@
 
 using Datadog.Trace.DuckTyping;
 
-namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.AspNetCore
+namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.AspNetCore.Proxies
 {
     [DuckCopy]
     internal struct TemplatePartStruct
