+++ tracer/src/Datadog.Trace/ClrProfiler/AutoInstrumentation/AspNetCore/Proxies/TemplateSegmentStruct.cs
@@ -1,4 +1,4 @@
-// <copyright file="IRouteData.cs" company="Datadog">
+// <copyright file="TemplateSegmentStruct.cs" company="Datadog">
 // Unless explicitly stated otherwise all files in this repository are licensed under the Apache 2 License.
 // This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2017 Datadog, Inc.
 // </copyright>
@@ -6,16 +6,15 @@
 #if NETFRAMEWORK
 #nullable enable
 
-using System.Collections.Generic;
+using System.Collections;
 using Datadog.Trace.DuckTyping;
 
-namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.AspNetCore
+namespace Datadog.Trace.ClrProfiler.AutoInstrumentation.AspNetCore.Proxies
 {
-    internal interface IRouteData : IDuckType
+    [DuckCopy]
+    internal struct TemplateSegmentStruct
     {
-        IEnumerable<object> Routers { get; }
-
-        IDictionary<string, object> Values { get; }
+        public IEnumerable Parts;
     }
 }
 #endif
